attached_assets/requirements_1753704956605.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-silero

python-dotenv
livekit-api
flask[async]
flask
flask-cors
uvicorn
livekit-plugins-google
livekit-plugins-noise-cancellation
mem0ai
duckduckgo-search
langchain_community
requests
tools